Refactored classes and namespace
