+++ src/main/java/org/right_brothers/agents/CoolingRackAgent.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.core.type.TypeReference;
 import org.right_brothers.agents.BaseAgent;
-import org.right_brothers.objects.ProcessedProduct;
+import org.right_brothers.bakery_objects.ProcessedProduct;
 import org.right_brothers.data.messages.ProductMessage;
 import org.right_brothers.utils.JsonConverter;
