+++ src/main/java/org/right_brothers/bakery_objects/BakedProduct.java
@@ -1,4 +1,4 @@
-package org.right_brothers.objects;
+package org.right_brothers.bakery_objects;
 
 import org.right_brothers.data.models.Step;
 import java.util.Vector;
